@@ -1,21 +1,32 @@
 {
 	"name": "chalk",
-	"version": "4.1.2",
+	"version": "5.6.0",
 	"description": "Terminal string styling done right",
 	"license": "MIT",
 	"repository": "chalk/chalk",
 	"funding": "https://github.com/chalk/chalk?sponsor=1",
-	"main": "source",
+	"type": "module",
+	"main": "./source/index.js",
+	"exports": "./source/index.js",
+	"imports": {
+		"#ansi-styles": "./source/vendor/ansi-styles/index.js",
+		"#supports-color": {
+			"node": "./source/vendor/supports-color/index.js",
+			"default": "./source/vendor/supports-color/browser.js"
+		}
+	},
+	"types": "./source/index.d.ts",
+	"sideEffects": false,
 	"engines": {
-		"node": ">=10"
+		"node": "^12.17.0 || ^14.13 || >=16.0.0"
 	},
 	"scripts": {
-		"test": "xo && nyc ava && tsd",
+		"test": "xo && c8 ava && tsd",
 		"bench": "matcha benchmark.js"
 	},
 	"files": [
 		"source",
-		"index.d.ts"
+		"!source/index.test-d.ts"
 	],
 	"keywords": [
 		"color",
@@ -25,7 +36,6 @@
 		"console",
 		"cli",
 		"string",
-		"str",
 		"ansi",
 		"style",
 		"styles",
@@ -40,29 +50,34 @@
 		"command-line",
 		"text"
 	],
-	"dependencies": {
-		"ansi-styles": "^4.1.0",
-		"supports-color": "^7.1.0"
-	},
 	"devDependencies": {
-		"ava": "^2.4.0",
-		"coveralls": "^3.0.7",
-		"execa": "^4.0.0",
-		"import-fresh": "^3.1.0",
+		"@types/node": "^16.11.10",
+		"ava": "^3.15.0",
+		"c8": "^7.10.0",
+		"color-convert": "^2.0.1",
+		"execa": "^6.0.0",
+		"log-update": "^5.0.0",
 		"matcha": "^0.7.0",
-		"nyc": "^15.0.0",
-		"resolve-from": "^5.0.0",
-		"tsd": "^0.7.4",
-		"xo": "^0.28.2"
+		"tsd": "^0.19.0",
+		"xo": "^0.57.0",
+		"yoctodelay": "^2.0.0"
 	},
 	"xo": {
 		"rules": {
 			"unicorn/prefer-string-slice": "off",
-			"unicorn/prefer-includes": "off",
-			"@typescript-eslint/member-ordering": "off",
-			"no-redeclare": "off",
-			"unicorn/string-content": "off",
-			"unicorn/better-regex": "off"
+			"@typescript-eslint/consistent-type-imports": "off",
+			"@typescript-eslint/consistent-type-exports": "off",
+			"@typescript-eslint/consistent-type-definitions": "off",
+			"unicorn/expiring-todo-comments": "off"
 		}
+	},
+	"c8": {
+		"reporter": [
+			"text",
+			"lcov"
+		],
+		"exclude": [
+			"source/vendor"
+		]
 	}
 }
