@@ -1,22 +1,24 @@
 {
   "name": "ipaddr.js",
   "description": "A library for manipulating IPv4 and IPv6 addresses in JavaScript.",
-  "version": "1.9.1",
+  "version": "2.2.0",
   "author": "whitequark <whitequark@whitequark.org>",
   "directories": {
     "lib": "./lib"
   },
   "dependencies": {},
   "devDependencies": {
-    "coffee-script": "~1.12.6",
-    "nodeunit": "^0.11.3",
-    "uglify-js": "~3.0.19"
+    "eslint": "^8.57.0",
+    "uglify-es": "*"
   },
   "scripts": {
-    "test": "cake build test"
+    "lint": "npx eslint lib",
+    "lintfix": "npx eslint --fix lib test",
+    "build": "npx uglifyjs --compress --mangle --wrap=window -o ipaddr.min.js lib/ipaddr.js",
+    "test": "node --test"
   },
   "files": [
-    "lib/",
+    "lib",
     "LICENSE",
     "ipaddr.min.js"
   ],
@@ -28,7 +30,7 @@
   "repository": "git://github.com/whitequark/ipaddr.js",
   "main": "./lib/ipaddr.js",
   "engines": {
-    "node": ">= 0.10"
+    "node": ">= 10"
   },
   "license": "MIT",
   "types": "./lib/ipaddr.js.d.ts"
