@@ -1,101 +1,63 @@
 {
   "name": "json-server",
-  "version": "0.17.1",
-  "description": "Get a full fake REST API with zero coding in less than 30 seconds",
-  "main": "./lib/server/index.js",
-  "bin": "./lib/cli/bin.js",
+  "version": "1.0.0-beta.3",
+  "description": "",
+  "type": "module",
+  "bin": {
+    "json-server": "lib/bin.js"
+  },
+  "types": "lib",
   "files": [
     "lib",
-    "public"
+    "views"
   ],
+  "engines": {
+    "node": ">=18.3"
+  },
   "scripts": {
-    "prepare": "husky install",
-    "test": "npm run build && cross-env NODE_ENV=test jest && npm run lint",
-    "start": "babel-node -- src/cli/bin db.json -r routes.json",
-    "lint": "eslint . --ignore-path .gitignore",
-    "fix": "npm run lint -- --fix",
-    "build": "babel src -d lib",
-    "toc": "markdown-toc -i README.md",
-    "postversion": "git push && git push --tags",
-    "prepublish": "npm test && npm run build"
+    "dev": "tsx watch src/bin.ts fixtures/db.json",
+    "build": "rm -rf lib && tsc",
+    "test": "node --import tsx/esm --test src/*.test.ts",
+    "lint": "eslint src",
+    "prepare": "husky",
+    "prepublishOnly": "npm run build"
   },
-  "dependencies": {
-    "body-parser": "^1.19.0",
-    "chalk": "^4.1.2",
-    "compression": "^1.7.4",
-    "connect-pause": "^0.1.1",
-    "cors": "^2.8.5",
-    "errorhandler": "^1.5.1",
-    "express": "^4.17.1",
-    "express-urlrewrite": "^1.4.0",
-    "json-parse-helpfulerror": "^1.0.3",
-    "lodash": "^4.17.21",
-    "lodash-id": "^0.14.1",
-    "lowdb": "^1.0.0",
-    "method-override": "^3.0.0",
-    "morgan": "^1.10.0",
-    "nanoid": "^3.1.23",
-    "please-upgrade-node": "^3.2.0",
-    "pluralize": "^8.0.0",
-    "server-destroy": "^1.0.1",
-    "yargs": "^17.0.1"
-  },
-  "devDependencies": {
-    "@babel/cli": "^7.12.1",
-    "@babel/core": "^7.12.3",
-    "@babel/node": "^7.12.6",
-    "@babel/preset-env": "^7.12.1",
-    "cross-env": "^7.0.2",
-    "eslint": "^7.13.0",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-config-standard": "^16.0.1",
-    "eslint-plugin-import": "^2.22.1",
-    "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-prettier": "^3.1.4",
-    "eslint-plugin-promise": "^4.2.1",
-    "eslint-plugin-standard": "^4.1.0",
-    "husky": "^6.0.0",
-    "jest": "^26.6.3",
-    "markdown-toc": "^1.2.0",
-    "mkdirp": "^1.0.4",
-    "npm-run-all": "^4.1.5",
-    "os-tmpdir": "^2.0.0",
-    "prettier": "^2.3.2",
-    "rimraf": "^3.0.2",
-    "server-ready": "^0.3.1",
-    "supertest": "^6.0.1",
-    "temp-write": "^4.0.0"
-  },
+  "keywords": [],
+  "author": "typicode <typicode@gmail.com>",
+  "license": "SEE LICENSE IN ./LICENSE",
   "repository": {
     "type": "git",
-    "url": "git://github.com/typicode/json-server.git"
+    "url": "git+https://github.com/typicode/json-server.git"
   },
-  "keywords": [
-    "JSON",
-    "server",
-    "fake",
-    "REST",
-    "API",
-    "prototyping",
-    "mock",
-    "mocking",
-    "test",
-    "testing",
-    "rest",
-    "data",
-    "dummy",
-    "sandbox"
-  ],
-  "author": "Typicode <typicode@gmail.com>",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/typicode/json-server/issues"
+  "devDependencies": {
+    "@eslint/js": "^9.11.0",
+    "@sindresorhus/tsconfig": "^6.0.0",
+    "@tailwindcss/typography": "^0.5.15",
+    "@types/node": "^22.5.5",
+    "concurrently": "^9.0.1",
+    "eslint": "^9.11.0",
+    "get-port": "^7.1.0",
+    "globals": "^15.9.0",
+    "husky": "^9.1.6",
+    "tempy": "^3.1.0",
+    "tsx": "^4.19.1",
+    "type-fest": "^4.26.1",
+    "typescript": "^5.6.2",
+    "typescript-eslint": "^8.6.0"
   },
-  "homepage": "https://github.com/typicode/json-server",
-  "engines": {
-    "node": ">=12"
-  },
-  "jest": {
-    "testURL": "http://localhost/"
+  "dependencies": {
+    "@tinyhttp/app": "^2.4.0",
+    "@tinyhttp/cors": "^2.0.1",
+    "@tinyhttp/logger": "^2.0.0",
+    "chalk": "^5.3.0",
+    "chokidar": "^4.0.1",
+    "dot-prop": "^9.0.0",
+    "eta": "^3.5.0",
+    "inflection": "^3.0.0",
+    "json5": "^2.2.3",
+    "lowdb": "^7.0.1",
+    "milliparsec": "^4.0.0",
+    "sirv": "^2.0.4",
+    "sort-on": "^6.1.0"
   }
 }
