@@ -1,80 +1,85 @@
 {
   "name": "lowdb",
-  "version": "1.0.0",
-  "description": "Small JSON database for Node, Electron and the browser. Powered by Lodash.",
+  "version": "7.0.1",
+  "description": "Tiny local JSON database for Node, Electron and the browser",
   "keywords": [
-    "flat",
-    "file",
-    "local",
     "database",
-    "storage",
-    "JSON",
-    "lodash",
-    "localStorage",
+    "db",
     "electron",
     "embed",
-    "embeddable"
+    "embedded",
+    "flat",
+    "JSON",
+    "local",
+    "localStorage",
+    "sessionStorage",
+    "browser",
+    "esm"
   ],
-  "main": "./lib/main.js",
-  "scripts": {
-    "test": "jest && npm run lint",
-    "lint": "eslint . --ignore-path .gitignore",
-    "fix": "npm run lint -- --fix",
-    "prepublishOnly": "npm run build && pkg-ok",
-    "build": "npm run build:lib && npm run build:dist",
-    "build:lib": "rimraf lib && babel src --out-dir lib && npm run mvAdapters",
-    "build:dist": "rimraf dist && webpack && webpack -p",
-    "mvAdapters": "rimraf adapters && mv lib/adapters .",
-    "precommit": "npm test"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git://github.com/typicode/lowdb.git"
-  },
-  "author": "Typicode <typicode@gmail.com>",
-  "license": "MIT",
+  "homepage": "https://github.com/typicode/lowdb#readme",
   "bugs": {
     "url": "https://github.com/typicode/lowdb/issues"
   },
-  "homepage": "https://github.com/typicode/lowdb",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/typicode/lowdb.git"
+  },
+  "funding": "https://github.com/sponsors/typicode",
+  "license": "MIT",
+  "author": "Typicode <typicode@gmail.com>",
+  "type": "module",
+  "exports": {
+    ".": "./lib/index.js",
+    "./node": "./lib/node.js",
+    "./browser": "./lib/browser.js"
+  },
+  "types": "./lib",
+  "typesVersions": {
+    "*": {
+      "node": [
+        "lib/node.d.ts"
+      ],
+      "browser": [
+        "lib/browser.d.ts"
+      ]
+    }
+  },
+  "files": [
+    "lib",
+    "!lib/examples/**/*",
+    "!lib/**/*.test.*"
+  ],
+  "scripts": {
+    "test": "node --import tsx/esm --test src/**/*.test.ts src/**/**/*.test.ts",
+    "lint": "eslint src --ext .ts --ignore-path .gitignore",
+    "build": "del-cli lib && tsc",
+    "prepublishOnly": "npm run build",
+    "postversion": "git push --follow-tags && npm publish",
+    "prepare": "husky install"
+  },
   "dependencies": {
-    "graceful-fs": "^4.1.3",
-    "is-promise": "^2.1.0",
-    "lodash": "4",
-    "pify": "^3.0.0",
-    "steno": "^0.4.1"
+    "steno": "^4.0.2"
   },
   "devDependencies": {
-    "babel-cli": "^6.2.0",
-    "babel-eslint": "^7.0.0",
-    "babel-jest": "^20.0.3",
-    "babel-loader": "^7.1.1",
-    "babel-polyfill": "^6.9.1",
-    "babel-preset-env": "^1.6.0",
-    "babel-register": "^6.9.0",
-    "delay": "^2.0.0",
-    "eslint": "^4.5.0",
-    "eslint-config-prettier": "^2.3.0",
-    "eslint-config-standard": "^10.2.1",
-    "eslint-plugin-import": "^2.6.1",
-    "eslint-plugin-node": "^5.1.0",
-    "eslint-plugin-prettier": "^2.1.2",
-    "eslint-plugin-promise": "^3.5.0",
-    "eslint-plugin-standard": "^3.0.1",
-    "husky": "^0.14.3",
-    "jest": "^20.0.4",
-    "lodash-id": "^0.14.0",
-    "mv": "^2.1.1",
-    "pkg-ok": "^1.0.1",
-    "prettier": "^1.5.2",
-    "ramda": "^0.24.1",
-    "regenerator-runtime": "^0.11.0",
-    "rimraf": "^2.5.4",
-    "sinon": "^3.2.1",
-    "tempfile": "^2.0.0",
-    "webpack": "^3.3.0"
+    "@commitlint/cli": "^18.4.3",
+    "@commitlint/config-conventional": "^18.4.3",
+    "@commitlint/prompt-cli": "^18.4.3",
+    "@sindresorhus/tsconfig": "^5.0.0",
+    "@types/express": "^4.17.21",
+    "@types/lodash": "^4.14.202",
+    "@types/node": "^20.10.5",
+    "@typicode/eslint-config": "^1.2.0",
+    "del-cli": "^5.1.0",
+    "eslint": "^8.56.0",
+    "express-async-handler": "^1.2.0",
+    "husky": "^8.0.3",
+    "lodash": "^4.17.21",
+    "tempy": "^3.1.0",
+    "ts-node": "^10.9.2",
+    "tsx": "^4.7.0",
+    "typescript": "^5.3.3"
   },
   "engines": {
-    "node": ">=4"
+    "node": ">=18"
   }
 }
