@@ -4,41 +4,73 @@
     "url": "http://github.com/broofa",
     "email": "robert@broofa.com"
   },
-  "bin": {
-    "mime": "cli.js"
-  },
+  "type": "module",
   "engines": {
-    "node": ">=4"
+    "node": ">=16"
   },
-  "contributors": [
-    {
-      "name": "Benjamin Thomas",
-      "url": "http://github.com/bentomas",
-      "email": "benjamin@benjaminthomas.org"
-    }
+  "main": "./dist/src/index.js",
+  "exports": {
+    ".": "./dist/src/index.js",
+    "./lite": "./dist/src/index_lite.js",
+    "./types/standard.js": "./dist/types/standard.js",
+    "./types/other.js": "./dist/types/other.js",
+    "./package.json": "./package.json"
+  },
+  "files": [
+    "bin",
+    "dist/src",
+    "dist/types",
+    "src",
+    "types"
   ],
+  "bin": {
+    "mime": "bin/cli.js"
+  },
+  "contributors": [],
   "description": "A comprehensive library for mime-type mapping",
   "license": "MIT",
-  "dependencies": {},
   "devDependencies": {
-    "github-release-notes": "0.13.1",
-    "mime-db": "1.31.0",
-    "mime-score": "1.1.0"
+    "@types/mime-db": "1.*",
+    "@types/mime-types": "2.1.4",
+    "@types/node": "20.14.10",
+    "@typescript-eslint/eslint-plugin": "7.15.0",
+    "@typescript-eslint/parser": "7.15.0",
+    "chalk": "5.3.0",
+    "mime-score": "2.0.4",
+    "mime-types": "2.1.35",
+    "prettier": "3.3.2",
+    "runmd": "1.3.9",
+    "standard-version": "9.5.0",
+    "typescript": "5.5.3"
   },
   "scripts": {
-    "prepare": "node src/build.js",
-    "changelog": "gren changelog --tags=all --generate --override",
-    "test": "node src/test.js"
+    "build": "npm run build:clean && tsc",
+    "build:clean": "rm -fr dist",
+    "build:types": "node dist/scripts/build.js",
+    "build:watch": "npm run build:clean && tsc --watch",
+    "lint": "prettier -c .",
+    "lint:fix": "prettier -w .",
+    "prepublishOnly": "npm run build && npm run build:types && npm test",
+    "release": "# `standard-version --dry-run` is the command you're after",
+    "test": "npm run build && node --test && ./test/exports_test.sh",
+    "test:watch": "clear && node --enable-source-maps --test --watch test"
   },
   "keywords": [
-    "util",
-    "mime"
+    "extension",
+    "file",
+    "mime",
+    "mime-db",
+    "mimetypes",
+    "util"
   ],
-  "main": "mime.js",
   "name": "mime",
   "repository": {
-    "url": "https://github.com/broofa/node-mime",
+    "url": "https://github.com/broofa/mime",
     "type": "git"
   },
-  "version": "1.6.0"
+  "version": "4.0.4",
+  "funding": [
+    "https://github.com/sponsors/broofa"
+  ],
+  "packageManager": "npm@10.8.1+sha256.b8807aebb9656758e2872fa6e7c564b506aa2faa9297439a478d471d2fe32483"
 }
