@@ -1,52 +1,30 @@
 {
   "name": "frontend",
-  "version": "0.0.0",
+  "version": "1.0.0",
+  "description": "This project was generated with [Angular CLI](https://github.com/angular/angular-cli) version 9.1.0.",
+  "main": "index.js",
   "scripts": {
-    "ng": "ng",
-    "start": "ng serve",
-    "build": "ng build",
-    "test": "ng test",
-    "lint": "ng lint",
-    "e2e": "ng e2e"
+    "test": "echo \"Error: no test specified\" && exit 1"
   },
-  "private": true,
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
   "dependencies": {
-<<<<<<< Updated upstream
-    "@angular/animations": "^15.1.2",
-    "@angular/cdk": "^15.1.1",
-    "@angular/common": "^15.1.2",
-    "@angular/compiler": "~9.1.0",
-    "@angular/core": "^15.1.2",
-    "@angular/forms": "^15.1.2",
-    "@angular/material": "^15.1.1",
-    "@angular/platform-browser": "^15.1.2",
-    "@angular/platform-browser-dynamic": "^15.1.2",
-    "@angular/router": "^15.1.2",
-=======
-    "@angular/animations": "^15.1.1",
-    "@angular/cdk": "^15.1.1",
-    "@angular/common": "^15.1.1",
-    "@angular/compiler": "~9.1.0",
-    "@angular/core": "^15.1.1",
-    "@angular/forms": "^15.1.1",
-    "@angular/material": "^15.1.1",
-    "@angular/platform-browser": "^15.1.1",
-    "@angular/platform-browser-dynamic": "^15.1.1",
-    "@angular/router": "^15.1.1",
->>>>>>> Stashed changes
-    "node": "^19.4.0",
-    "rxjs": "~6.5.4",
-    "tslib": "^1.10.0",
-    "zone.js": "~0.10.2"
+    "@angular/animations": "^18.2.11",
+    "@angular/cdk": "^18.2.12",
+    "@angular/cli": "^18.2.11",
+    "@angular/common": "^18.2.11",
+    "@angular/compiler": "^18.2.11",
+    "@angular/core": "^18.2.11",
+    "@angular/forms": "^18.2.11",
+    "@angular/material": "^18.2.12",
+    "@angular/platform-browser": "^18.2.11",
+    "@angular/platform-browser-dynamic": "^18.2.11",
+    "@angular/router": "^18.2.11"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^15.1.2",
-    "@angular/cli": "~9.1.0",
-    "@angular/compiler-cli": "~9.1.0",
-    "@angular/language-service": "~9.1.0",
-    "@types/node": "^12.11.1",
-    "ts-node": "~8.3.0",
-    "tslint": "~6.1.0",
-    "typescript": "~3.8.3"
+    "@angular-devkit/build-angular": "^18.2.11",
+    "@types/node": "^22.9.0",
+    "typescript": "5.5"
   }
 }
