@@ -1,31 +1,31 @@
-import { Product } from './../product.model';
-import { DataSource } from '@angular/cdk/collections';
-import { MatPaginator } from '@angular/material/paginator';
-import { MatSort } from '@angular/material/sort';
-import { map } from 'rxjs/operators';
-import { Observable, of as observableOf, merge } from 'rxjs';
+import { Product } from "./../product.model";
+import { DataSource } from "@angular/cdk/collections";
+import { MatPaginator } from "@angular/material/paginator";
+import { MatSort } from "@angular/material/sort";
+import { map } from "rxjs/operators";
+import { Observable, of as observableOf, merge } from "rxjs";
 
 const EXAMPLE_DATA: Product[] = [
-  {id: 1, name: 'Hydrogen', price: 9.99},
-  {id: 2, name: 'Helium', price: 9.99},
-  {id: 3, name: 'Lithium', price: 9.99},
-  {id: 4, name: 'Beryllium', price: 9.99},
-  {id: 5, name: 'Boron', price: 9.99},
-  {id: 6, name: 'Carbon', price: 9.99},
-  {id: 7, name: 'Nitrogen', price: 9.99},
-  {id: 8, name: 'Oxygen', price: 9.99},
-  {id: 9, name: 'Fluorine', price: 9.99},
-  {id: 10, name: 'Neon', price: 9.99},
-  {id: 11, name: 'Sodium', price: 9.99},
-  {id: 12, name: 'Magnesium', price: 9.99},
-  {id: 13, name: 'Aluminum', price: 9.99},
-  {id: 14, name: 'Silicon', price: 9.99},
-  {id: 15, name: 'Phosphorus', price: 9.99},
-  {id: 16, name: 'Sulfur', price: 9.99},
-  {id: 17, name: 'Chlorine', price: 9.99},
-  {id: 18, name: 'Argon', price: 9.99},
-  {id: 19, name: 'Potassium', price: 9.99},
-  {id: 20, name: 'Calcium', price: 9.99},
+  { id: 1, name: "Hydrogen", price: 9.99 },
+  { id: 2, name: "Helium", price: 9.99 },
+  { id: 3, name: "Lithium", price: 9.99 },
+  { id: 4, name: "Beryllium", price: 9.99 },
+  { id: 5, name: "Boron", price: 9.99 },
+  { id: 6, name: "Carbon", price: 9.99 },
+  { id: 7, name: "Nitrogen", price: 9.99 },
+  { id: 8, name: "Oxygen", price: 9.99 },
+  { id: 9, name: "Fluorine", price: 9.99 },
+  { id: 10, name: "Neon", price: 9.99 },
+  { id: 11, name: "Sodium", price: 9.99 },
+  { id: 12, name: "Magnesium", price: 9.99 },
+  { id: 13, name: "Aluminum", price: 9.99 },
+  { id: 14, name: "Silicon", price: 9.99 },
+  { id: 15, name: "Phosphorus", price: 9.99 },
+  { id: 16, name: "Sulfur", price: 9.99 },
+  { id: 17, name: "Chlorine", price: 9.99 },
+  { id: 18, name: "Argon", price: 9.99 },
+  { id: 19, name: "Potassium", price: 9.99 },
+  { id: 20, name: "Calcium", price: 9.99 },
 ];
 
 /**
@@ -52,13 +52,15 @@ export class ProductRead2DataSource extends DataSource<Product> {
     // stream for the data-table to consume.
     const dataMutations = [
       observableOf(this.data),
-      this.paginator.page,
-      this.sort.sortChange
+      this.paginator.page.asObservable(),
+      this.sort.sortChange.asObservable(),
     ];
 
-    return merge(...dataMutations).pipe(map(() => {
-      return this.getPagedData(this.getSortedData([...this.data]));
-    }));
+    return merge(...dataMutations).pipe(
+      map(() => {
+        return this.getPagedData(this.getSortedData([...this.data]));
+      })
+    );
   }
 
   /**
@@ -81,16 +83,19 @@ export class ProductRead2DataSource extends DataSource<Product> {
    * this would be replaced by requesting the appropriate data from the server.
    */
   private getSortedData(data: Product[]) {
-    if (!this.sort.active || this.sort.direction === '') {
+    if (!this.sort.active || this.sort.direction === "") {
       return data;
     }
 
     return data.sort((a, b) => {
-      const isAsc = this.sort.direction === 'asc';
+      const isAsc = this.sort.direction === "asc";
       switch (this.sort.active) {
-        case 'name': return compare(a.name, b.name, isAsc);
-        case 'id': return compare(+a.id, +b.id, isAsc);
-        default: return 0;
+        case "name":
+          return compare(a.name, b.name, isAsc);
+        case "id":
+          return compare(+a.id, +b.id, isAsc);
+        default:
+          return 0;
       }
     });
   }
